Raise an exception when incorrect data is passed to the Notifications::NewFollower service
