+++ app/services/notifications/new_follower/follow_data.rb
@@ -1,0 +1,16 @@
+require "dry-types"
+require "dry-struct"
+
+module Notifications
+  module NewFollower
+    module Types
+      include Dry::Types.module
+    end
+
+    class FollowData < Dry::Struct
+      attribute :followable_id, Types::Strict::Integer
+      attribute :followable_type, Types::Strict::String.enum("User", "Organization")
+      attribute :follower_id, Types::Strict::Integer
+    end
+  end
+end
